+++ circuit-codegen/src/main/kotlin/com/slack/circuit/codegen/CircuitSymbolProcessorProvider.kt
@@ -62,29 +62,19 @@ public class CircuitSymbolProcessorProvider : SymbolProcessorProvider {
 private class CircuitSymbolProcessor(
   private val logger: KSPLogger,
   private val codeGenerator: CodeGenerator,
-  private val options: Map<String, String>,
+  options: Map<String, String>,
   private val platforms: List<PlatformInfo>,
 ) : SymbolProcessor {
 
-  private val lenient = options["circuit.codegen.lenient"]?.toBoolean() ?: false
+  private val options = CircuitOptions.load(options, logger)
+  private val mode = this.options.mode
 
   override fun process(resolver: Resolver): List<KSAnnotated> {
+    if (options === CircuitOptions.UNKNOWN) return emptyList()
     val symbols = CircuitSymbols.create(resolver) ?: return emptyList()
-    val codegenMode =
-      options[CircuitNames.CIRCUIT_CODEGEN_MODE].let { mode ->
-        if (mode == null) {
-          CodegenMode.ANVIL
-        } else {
-          CodegenMode.entries.find { it.name.equals(mode, ignoreCase = true) }
-            ?: run {
-              logger.error("Unrecognised option for codegen mode \"$mode\".")
-              return emptyList()
-            }
-        }
-      }
 
-    if (!codegenMode.supportsPlatforms(platforms)) {
-      logger.error("Unsupported platforms for codegen mode ${codegenMode.name}. $platforms")
+    if (!options.mode.supportsPlatforms(platforms)) {
+      logger.error("Unsupported platforms for codegen mode ${mode.name}. $platforms")
       return emptyList()
     }
 
@@ -93,10 +83,10 @@ private class CircuitSymbolProcessor(
       .forEach { annotatedElement ->
         when (annotatedElement) {
           is KSClassDeclaration ->
-            generateFactory(annotatedElement, InstantiationType.CLASS, symbols, codegenMode)
+            generateFactory(annotatedElement, InstantiationType.CLASS, symbols)
 
           is KSFunctionDeclaration ->
-            generateFactory(annotatedElement, InstantiationType.FUNCTION, symbols, codegenMode)
+            generateFactory(annotatedElement, InstantiationType.FUNCTION, symbols)
 
           else ->
             logger.error(
@@ -113,7 +103,6 @@ private class CircuitSymbolProcessor(
     annotatedElement: KSAnnotated,
     instantiationType: InstantiationType,
     symbols: CircuitSymbols,
-    codegenMode: CodegenMode,
   ) {
     val circuitInjectAnnotation =
       annotatedElement.getKSAnnotationsWithLeniency(CircuitNames.CIRCUIT_INJECT_ANNOTATION).single()
@@ -121,7 +110,7 @@ private class CircuitSymbolProcessor(
     // If we annotated a class, check that the class isn't using assisted inject. If so, error and
     // return
     if (instantiationType == InstantiationType.CLASS) {
-      (annotatedElement as KSClassDeclaration).checkForAssistedInjection(codegenMode) {
+      (annotatedElement as KSClassDeclaration).checkForAssistedInjection(mode) {
         return
       }
     }
@@ -139,7 +128,7 @@ private class CircuitSymbolProcessor(
         screenKSType,
         instantiationType,
         logger,
-        codegenMode = codegenMode,
+        mode = mode,
       ) ?: return
 
     val className =
@@ -157,7 +146,7 @@ private class CircuitSymbolProcessor(
         // Add the `@Inject` annotation to the appropriate place
         val constructorBuilder =
           FunSpec.constructorBuilder().addParameters(factoryData.constructorParams)
-        codegenMode.addInjectAnnotation(this, constructorBuilder)
+        mode.addInjectAnnotation(this, constructorBuilder, options)
         if (
           constructorBuilder.annotations.isNotEmpty() || constructorBuilder.parameters.isNotEmpty()
         ) {
@@ -174,10 +163,10 @@ private class CircuitSymbolProcessor(
           }
         }
 
-        codegenMode.annotateFactory(builder = this, scope = scope)
-        if (topLevelClass != null && codegenMode.originAnnotation != null) {
+        mode.annotateFactory(builder = this, scope = scope)
+        if (topLevelClass != null && mode.originAnnotation != null) {
           addAnnotation(
-            AnnotationSpec.builder(codegenMode.originAnnotation)
+            AnnotationSpec.builder(mode.originAnnotation)
               .addMember("%T::class", topLevelClass)
               .build()
           )
@@ -209,7 +198,7 @@ private class CircuitSymbolProcessor(
       )
 
     val additionalType =
-      codegenMode.produceAdditionalTypeSpec(
+      mode.produceAdditionalTypeSpec(
         factoryType = factoryData.factoryType,
         factory = ClassName(factoryData.packageName, className + CircuitNames.FACTORY),
         scope = scope,
@@ -233,17 +222,16 @@ private class CircuitSymbolProcessor(
   }
 
   private inline fun KSClassDeclaration.checkForAssistedInjection(
-    codegenMode: CodegenMode,
+    mode: CodegenMode,
     exit: () -> Nothing,
   ) {
-    val assistedInjectClassName = codegenMode.runtime.assistedInject ?: return
+    val assistedInjectClassName = mode.runtime.assistedInject ?: return
     // Check for an AssistedInject constructor
     if (findConstructorAnnotatedWith(assistedInjectClassName) != null) {
       val assistedFactory =
         declarations.filterIsInstance<KSClassDeclaration>().find { nestedClass ->
-          codegenMode.runtime.assistedFactory?.let {
-            nestedClass.isAnnotationPresentWithLeniency(it)
-          } == true
+          mode.runtime.assistedFactory?.let { nestedClass.isAnnotationPresentWithLeniency(it) } ==
+            true
         }
       val suffix =
         if (assistedFactory != null) " (${assistedFactory.qualifiedName?.asString()})" else ""
@@ -263,7 +251,7 @@ private class CircuitSymbolProcessor(
     annotation: ClassName
   ): Sequence<KSAnnotation> {
     val simpleName = annotation.simpleName
-    return if (lenient) {
+    return if (options.lenient) {
       annotations.filter { it.shortName.asString() == simpleName }
     } else {
       val qualifiedName = annotation.canonicalName
@@ -291,7 +279,7 @@ private class CircuitSymbolProcessor(
     screenKSType: KSType,
     instantiationType: InstantiationType,
     logger: KSPLogger,
-    codegenMode: CodegenMode,
+    mode: CodegenMode,
   ): FactoryData? {
     val className: String
     val packageName: String
@@ -320,7 +308,7 @@ private class CircuitSymbolProcessor(
             logger,
             screenKSType,
             factoryType == FactoryType.PRESENTER,
-            includeParameterNames = codegenMode != KOTLIN_INJECT_ANVIL,
+            includeParameterNames = mode != KOTLIN_INJECT_ANVIL,
           )
 
         // Check we have a return type
@@ -422,27 +410,26 @@ private class CircuitSymbolProcessor(
         }
         val injectableConstructor by
           lazy(NONE) {
-            declaration.findConstructorAnnotatedWith(codegenMode.runtime.inject)
+            declaration.findConstructorAnnotatedWith(mode.runtime.inject(options))
               ?: declaration.primaryConstructor
           }
         val assistedKSParams by
           lazy(NONE) {
             injectableConstructor
               ?.parameters
-              ?.filter { it.isAnnotationPresentWithLeniency(codegenMode.runtime.assisted) }
+              ?.filter { it.isAnnotationPresentWithLeniency(mode.runtime.assisted) }
               .orEmpty()
           }
 
         val isAssisted =
           assistedKSParams.isNotEmpty() ||
-            codegenMode.runtime.assistedFactory?.let {
-              declaration.isAnnotationPresentWithLeniency(it)
-            } == true
+            mode.runtime.assistedFactory?.let { declaration.isAnnotationPresentWithLeniency(it) } ==
+              true
 
         val creatorOrConstructor: KSFunctionDeclaration?
         val targetClass: KSClassDeclaration
         if (isAssisted) {
-          if (codegenMode == KOTLIN_INJECT_ANVIL) {
+          if (mode == KOTLIN_INJECT_ANVIL) {
             creatorOrConstructor = injectableConstructor
             targetClass = declaration
           } else {
@@ -459,9 +446,9 @@ private class CircuitSymbolProcessor(
         }
         val useProvider =
           !isAssisted &&
-            codegenMode
-              .filterValidInjectionSites(listOfNotNull(creatorOrConstructor, declaration))
-              .any { it.isAnnotationPresentWithLeniency(codegenMode.runtime.inject) }
+            mode.filterValidInjectionSites(listOfNotNull(creatorOrConstructor, declaration)).any {
+              it.isAnnotationPresentWithLeniency(mode.runtime.inject(options))
+            }
         className = targetClass.simpleName.getShortName()
         packageName = targetClass.packageName.asString()
         factoryType =
@@ -500,7 +487,7 @@ private class CircuitSymbolProcessor(
               logger,
               screenKSType,
               allowNavigator = factoryType == FactoryType.PRESENTER,
-              includeParameterNames = codegenMode != KOTLIN_INJECT_ANVIL,
+              includeParameterNames = mode != KOTLIN_INJECT_ANVIL,
             )
           }
         codeBlock =
@@ -509,14 +496,14 @@ private class CircuitSymbolProcessor(
             constructorParams.add(
               ParameterSpec.builder(
                   "provider",
-                  codegenMode.runtime.asProvider(targetClass.toClassName()),
+                  mode.runtime.asProvider(targetClass.toClassName(), options),
                 )
                 .build()
             )
-            codegenMode.runtime.getProviderBlock(CodeBlock.of("provider"))
+            mode.runtime.getProviderBlock(CodeBlock.of("provider"))
           } else if (isAssisted) {
             // Inject the target class's assisted factory that we'll call its create() on.
-            when (codegenMode) {
+            when (mode) {
               KOTLIN_INJECT_ANVIL -> {
                 val factoryLambda =
                   LambdaTypeName.get(