+++ circuit-codegen/src/main/kotlin/com/slack/circuit/codegen/CodegenMode.kt
@@ -19,6 +19,19 @@ import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeSpec
 
 internal enum class CodegenMode {
+  UNKNOWN {
+    override fun supportsPlatforms(platforms: List<PlatformInfo>): Boolean {
+      return false
+    }
+
+    override fun addInjectAnnotation(
+      classBuilder: TypeSpec.Builder,
+      constructorBuilder: FunSpec.Builder,
+      options: CircuitOptions,
+    ) {
+      error("Unknown codegen mode, should not be called")
+    }
+  },
   /**
    * The Anvil Codegen mode
    *
@@ -51,8 +64,9 @@ internal enum class CodegenMode {
     override fun addInjectAnnotation(
       classBuilder: TypeSpec.Builder,
       constructorBuilder: FunSpec.Builder,
+      options: CircuitOptions,
     ) {
-      constructorBuilder.addAnnotation(runtime.inject)
+      constructorBuilder.addAnnotation(runtime.inject(options))
     }
   },
 
@@ -138,8 +152,9 @@ internal enum class CodegenMode {
     override fun addInjectAnnotation(
       classBuilder: TypeSpec.Builder,
       constructorBuilder: FunSpec.Builder,
+      options: CircuitOptions,
     ) {
-      constructorBuilder.addAnnotation(runtime.inject)
+      constructorBuilder.addAnnotation(runtime.inject(options))
     }
   },
 
@@ -179,8 +194,9 @@ internal enum class CodegenMode {
     override fun addInjectAnnotation(
       classBuilder: TypeSpec.Builder,
       constructorBuilder: FunSpec.Builder,
+      options: CircuitOptions,
     ) {
-      classBuilder.addAnnotation(runtime.inject)
+      classBuilder.addAnnotation(runtime.inject(options))
     }
 
     override fun filterValidInjectionSites(
@@ -225,8 +241,9 @@ internal enum class CodegenMode {
     override fun addInjectAnnotation(
       classBuilder: TypeSpec.Builder,
       constructorBuilder: FunSpec.Builder,
+      options: CircuitOptions,
     ) {
-      classBuilder.addAnnotation(runtime.inject)
+      classBuilder.addAnnotation(runtime.inject(options))
     }
 
     override fun filterValidInjectionSites(
@@ -255,6 +272,7 @@ internal enum class CodegenMode {
   abstract fun addInjectAnnotation(
     classBuilder: TypeSpec.Builder,
     constructorBuilder: FunSpec.Builder,
+    options: CircuitOptions,
   )
 
   /** Filters the candidates for @Inject annotation placement. */
@@ -263,23 +281,27 @@ internal enum class CodegenMode {
   ): Collection<KSDeclaration> = candidates.filterIsInstance<KSFunctionDeclaration>()
 
   sealed interface InjectionRuntime {
-    val inject: ClassName
+    fun inject(options: CircuitOptions): ClassName
+
     val assisted: ClassName
     val assistedInject: ClassName?
     val assistedFactory: ClassName?
 
-    fun asProvider(providedType: TypeName): TypeName
+    fun asProvider(providedType: TypeName, options: CircuitOptions): TypeName
 
     fun getProviderBlock(provider: CodeBlock): CodeBlock
 
     data object Jakarta : InjectionRuntime {
-      override val inject: ClassName = CircuitNames.INJECT
+      override fun inject(options: CircuitOptions): ClassName =
+        if (options.useJavax) CircuitNames.INJECT_JAVAX else CircuitNames.INJECT
+
       override val assisted: ClassName = CircuitNames.ASSISTED
       override val assistedInject: ClassName = CircuitNames.ASSISTED_INJECT
       override val assistedFactory: ClassName = CircuitNames.ASSISTED_FACTORY
 
-      override fun asProvider(providedType: TypeName): TypeName {
-        return CircuitNames.PROVIDER.parameterizedBy(providedType)
+      override fun asProvider(providedType: TypeName, options: CircuitOptions): TypeName {
+        val className = if (options.useJavax) CircuitNames.PROVIDER_JAVAX else CircuitNames.PROVIDER
+        return className.parameterizedBy(providedType)
       }
 
       override fun getProviderBlock(provider: CodeBlock): CodeBlock {
@@ -288,12 +310,13 @@ internal enum class CodegenMode {
     }
 
     data object KotlinInject : InjectionRuntime {
-      override val inject: ClassName = CircuitNames.KotlinInject.INJECT
+      override fun inject(options: CircuitOptions) = CircuitNames.KotlinInject.INJECT
+
       override val assisted: ClassName = CircuitNames.KotlinInject.ASSISTED
       override val assistedInject: ClassName? = null
       override val assistedFactory: ClassName? = null // It has no annotation
 
-      override fun asProvider(providedType: TypeName): TypeName {
+      override fun asProvider(providedType: TypeName, options: CircuitOptions): TypeName {
         return LambdaTypeName.get(returnType = providedType)
       }
 
@@ -303,12 +326,13 @@ internal enum class CodegenMode {
     }
 
     data object Metro : InjectionRuntime {
-      override val inject: ClassName = CircuitNames.Metro.INJECT
+      override fun inject(options: CircuitOptions) = CircuitNames.Metro.INJECT
+
       override val assisted: ClassName = CircuitNames.Metro.ASSISTED
       override val assistedInject: ClassName? = null
       override val assistedFactory: ClassName = CircuitNames.Metro.ASSISTED_FACTORY
 
-      override fun asProvider(providedType: TypeName): TypeName {
+      override fun asProvider(providedType: TypeName, options: CircuitOptions): TypeName {
         return CircuitNames.Metro.PROVIDER.parameterizedBy(providedType)
       }